+++ frmTray.cs
@@ -44,7 +44,7 @@ namespace NUTty_UPS_Client
 
             this.txtUPSStatistics.Text = NUT_Processor.ParseNUTOutput(NUT_poller.PollNUTServer("192.168.253.6", 3493));
             lblUPSConnection.Text = "Connected";
-            Tuple<string, int, int> UPSBatteryStatus = NUT_Processor.GetBatteryStatus();
+            Tuple<string, double, int> UPSBatteryStatus = NUT_Processor.GetBatteryStatus();
             lblTimeRemaining.Text = UPSBatteryStatus.Item1;
             lblBatteryPercentage.Text = Convert.ToString(UPSBatteryStatus.Item2) + "%";
